ftdi: drop the libftdi C backend
With ftdi-nusb 0.3.0 the pure-Rust backend covers everything the
libftdi1 one did (flashrom-compatible CS masks, auxiliary and
high-byte pin state, batched single-USB-write transactions), so the
last hand-rolled MPSSE implementation and the C library dependency
can go.

The `ftdi` feature now selects the ftdi-nusb backend directly and the
`ftdi-native` / `all-programmers-native` split disappears; the backend
files lose their rsftdi_ prefix. CI, the Nix flake, and the README no
longer need libftdi1.
